pyproject.toml
[tool.poetry]
name = "pynethernet"
version = "0.1.0"
description = "Provides an API allowing users to interact with Minecraft: Bedrock Edition realms and servers via the NetherNet WebRTC protocol."
authors = ["DJ Stomp <85457381+DJStompZone@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
websockets = "^12.0"
aiortc = "^1.9.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pycryptodome = "^3.20.0"
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
